Fix : multipart/form 이 PATCH를 지원하지 않아 POST로 변경, 일부 로직 수정

+++ src/main/java/com/study/codingswamp/common/file/FireBaseFileStore.java
@@ -37,6 +37,9 @@ public class FireBaseFileStore implements FileStore{
 
     @Override
     public void deleteFile(String imageUrl) {
+        if (imageUrl.equals(defaultImageUrl)) {
+            return;
+        }
         String fileName = deleteFileNameExtract(imageUrl);
         Bucket bucket = StorageClient.getInstance().bucket(firebaseBucket);
         Blob blob = bucket.get(fileName);
